# stdenv: re-add meta.repository (resurrected)