--- a/README.md
+++ b/README.md
@@ -99,3 +99,5 @@ scripts, NixOS modules, etc.). It also might not apply to patches
 included in Nixpkgs, which may be derivative works of the packages to
 which they apply. The aforementioned artifacts are all covered by the
 licenses of the respective packages.
+
+<!-- re-trigger CI, this change will be removed before merging -->