@nixos-discourse 2024-04-06
This pull request has been mentioned on **NixOS Discourse**. There might be relevant details there:

https://discourse.nixos.org/t/when-should-meta-attributes-be-computed/42833/1