@annafleed 2024-04-09
> I think the approach makes sense, I would try to see if `builtins.tryEval` is usable and has no performance impact.

wrapping in `builtins.tryEval` has a negligible performance impact, so should be added IMO.

median of 11 runs  (only really matters for cpu time, everything else is stable i think).

using change 253c0db644cd27343edb39f243a360b2be2e2257 and x86_64-linux check [meta tests](https://github.com/NixOS/ofborg?tab=readme-ov-file#running-meta-checks-locally)

baseline is change 253c0db644cd27343edb39f243a360b2be2e2257, prlist11 is with this PR, prlist15 and prlist16 use tryEval

|          name          |   baseline    | prlist11 / baseline | prlist15 / baseline | prlist16 / baseline |
|------------------------|--------------:|--------------------:|--------------------:|--------------------:|
| cpuTime                | 127.11        | 0.9885              | 0.9994              | 1.0034              |
| envs.bytes             | 6551424928.0  | 1.0011              | 1.0013              | 1.0013              |
| envs.elements          | 344611966.0   | 1.0011              | 1.0013              | 1.0013              |
| envs.number            | 237158075.0   | 1.0011              | 1.0013              | 1.0013              |
| gc.heapSize            | 19734130688.0 | 1.0                 | 0.9991              | 0.9991              |
| gc.totalBytes          | 40074163200.0 | 1.0019              | 1.0022              | 1.0022              |
| list.bytes             | 767523792.0   | 1.0001              | 1.0001              | 1.0001              |
| list.concats           | 17306647.0    | 1.0                 | 1.0                 | 1.0                 |
| list.elements          | 95940474.0    | 1.0001              | 1.0001              | 1.0001              |
| nrAvoided              | 274572669.0   | 1.001               | 1.001               | 1.001               |
| nrFunctionCalls        | 218766710.0   | 1.001               | 1.001               | 1.0013              |
| nrLookups              | 109634172.0   | 1.0039              | 1.0058              | 1.0058              |
| nrOpUpdateValuesCopied | 606519203.0   | 1.0012              | 1.0012              | 1.0012              |
| nrOpUpdates            | 27664906.0    | 1.0055              | 1.0055              | 1.0055              |
| nrPrimOpCalls          | 122442293.0   | 1.0007              | 1.0013              | 1.0013              |
| nrThunks               | 334642079.0   | 1.0024              | 1.0028              | 1.0028              |
| sets.bytes             | 13762198032.0 | 1.0022              | 1.0024              | 1.0024              |
| sets.elements          | 809415832.0   | 1.0019              | 1.0021              | 1.0021              |
| sets.number            | 50721545.0    | 1.0066              | 1.008               | 1.008               |
| sizes.Attr             | 16.0          | 1.0                 | 1.0                 | 1.0                 |
| sizes.Bindings         | 16.0          | 1.0                 | 1.0                 | 1.0                 |
| sizes.Env              | 16.0          | 1.0                 | 1.0                 | 1.0                 |
| sizes.Value            | 24.0          | 1.0                 | 1.0                 | 1.0                 |
| symbols.bytes          | 2341038.0     | 1.0                 | 1.0                 | 1.0                 |
| symbols.number         | 168577.0      | 1.0                 | 1.0                 | 1.0                 |
| values.bytes           | 10281315384.0 | 1.0023              | 1.0028              | 1.0028              |
| values.number          | 428388141.0   | 1.0023              | 1.0028              | 1.0028              |

|   name   | geomean |
|----------|--------:|
| prlist11 | 1.0009  |
| prlist15 | 1.0015  |
| prlist16 | 1.0017  |


then running with `(<pkg>.meta.repository or null)` for all x64-linux packages also has negligible performance impact.
baseline is this PR, prlist15 and prlist16 are variations of `tryEval`
```console
NIX_SHOW_STATS=1 NIX_SHOW_STATS_PATH=results/baseline-stats$i.json nix eval -f - --show-trace <<EOF
with import ./.{}; ((import ./out-paths.nix) pkgs)
EOF
```

|          name          |   baseline   | prlist15 / baseline | prlist16 / baseline |
|------------------------|-------------:|--------------------:|--------------------:|
| cpuTime                | 18.26        | 1.0004              | 1.0042              |
| envs.bytes             | 1031048184.0 | 1.0015              | 1.0015              |
| envs.elements          | 60505417.0   | 1.0011              | 1.0011              |
| envs.number            | 34187803.0   | 1.0019              | 1.0019              |
| gc.heapSize            | 4699906048.0 | 0.9786              | 0.9786              |
| gc.totalBytes          | 7549011504.0 | 1.0015              | 1.0015              |
| list.bytes             | 84133128.0   | 1.0                 | 1.0                 |
| list.concats           | 1534723.0    | 1.0                 | 1.0                 |
| list.elements          | 10516641.0   | 1.0                 | 1.0                 |
| nrAvoided              | 38361747.0   | 1.0                 | 1.0                 |
| nrFunctionCalls        | 29846130.0   | 1.0                 | 1.0022              |
| nrLookups              | 17079015.0   | 1.0113              | 1.0113              |
| nrOpUpdateValuesCopied | 137632128.0  | 1.0                 | 1.0                 |
| nrOpUpdates            | 6493171.0    | 1.0                 | 1.0                 |
| nrPrimOpCalls          | 12235438.0   | 1.0053              | 1.0053              |
| nrThunks               | 61452402.0   | 1.0021              | 1.0021              |
| sets.bytes             | 3186938304.0 | 1.001               | 1.001               |
| sets.elements          | 186209984.0  | 1.0007              | 1.0007              |
| sets.number            | 12973660.0   | 1.005               | 1.005               |
| sizes.Attr             | 16.0         | 1.0                 | 1.0                 |
| sizes.Bindings         | 16.0         | 1.0                 | 1.0                 |
| sizes.Env              | 16.0         | 1.0                 | 1.0                 |
| sizes.Value            | 24.0         | 1.0                 | 1.0                 |
| symbols.bytes          | 1662469.0    | 1.0                 | 1.0                 |
| symbols.number         | 128122.0     | 1.0                 | 1.0                 |
| values.bytes           | 1908011976.0 | 1.0024              | 1.0024              |
| values.number          | 79500499.0   | 1.0024              | 1.0024              |

|   name   | geomean |
|----------|--------:|
| prlist15 | 1.0005  |
| prlist16 | 1.0008  |



<details><summary>prlist11.diff</summary>

```diff
diff --git a/pkgs/stdenv/generic/check-meta.nix b/pkgs/stdenv/generic/check-meta.nix
index 1cd1ae6dd72e..f1f486b115a8 100644
--- a/pkgs/stdenv/generic/check-meta.nix
+++ b/pkgs/stdenv/generic/check-meta.nix
@@ -10,7 +10,9 @@ let
     concatMapStrings
     concatMapStringsSep
     concatStrings
+    filter
     findFirst
+    head
     isDerivation
     length
     concatMap
@@ -39,6 +41,11 @@ let
     toPretty
   ;
 
+  unlist = list:
+  if length list == 1
+  then head list
+  else list;
+
   # If we're in hydra, we can dispense with the more verbose error
   # messages and make problems easier to spot.
   inHydra = config.inHydra or false;
@@ -303,6 +310,10 @@ let
       (listOf str)
       str
     ];
+    repository = union [
+      (listOf str)
+      str
+    ];
     downloadPage = str;
     changelog = union [
       (listOf str)
@@ -452,8 +463,14 @@ let
     let
       outputs = attrs.outputs or [ "out" ];
       hasOutput = out: builtins.elem out outputs;
-    in
-    {
+    in {
+      repository =
+        if attrs ? src.meta.homepage
+        then attrs.src.meta.homepage
+        else if attrs ? srcs && isList attrs.srcs
+        then unlist (map (src: src.meta.homepage) (filter (src: src ? meta.homepage) attrs.srcs))
+        else [];
+
       # `name` derivation attribute includes cross-compilation cruft,
       # is under assert, and is sanitized.
       # Let's have a clean always accessible version here.
```
</details>
<details><summary>prlist15.diff</summary>

```diff
diff --git a/pkgs/stdenv/generic/check-meta.nix b/pkgs/stdenv/generic/check-meta.nix
index bcb2ca249ddf..18764e52a319 100644
--- a/pkgs/stdenv/generic/check-meta.nix
+++ b/pkgs/stdenv/generic/check-meta.nix
@@ -10,7 +10,9 @@ let
     concatMapStrings
     concatMapStringsSep
     concatStrings
+    filter
     findFirst
+    head
     isDerivation
     length
     concatMap
@@ -39,6 +41,11 @@ let
     toPretty
   ;
 
+  unlist = list:
+  if length list == 1
+  then head list
+  else list;
+
   # If we're in hydra, we can dispense with the more verbose error
   # messages and make problems easier to spot.
   inHydra = config.inHydra or false;
@@ -303,6 +310,10 @@ let
       (listOf str)
       str
     ];
+    repository = union [
+      (listOf str)
+      str
+    ];
     downloadPage = str;
     changelog = union [
       (listOf str)
@@ -454,8 +465,15 @@ let
     let
       outputs = attrs.outputs or [ "out" ];
       hasOutput = out: builtins.elem out outputs;
-    in
-    {
+    in {
+      repository = let res = builtins.tryEval (
+        if attrs ? src.meta.homepage
+        then attrs.src.meta.homepage
+        else if attrs ? srcs && isList attrs.srcs
+        then unlist (map (src: src.meta.homepage) (filter (src: src ? meta.homepage) attrs.srcs))
+        else []);
+        in if res.success then res.value else [];
+
       # `name` derivation attribute includes cross-compilation cruft,
       # is under assert, and is sanitized.
       # Let's have a clean always accessible version here.
```
</details>
<details><summary>prlist16.diff</summary>

```diff
diff --git a/pkgs/stdenv/generic/check-meta.nix b/pkgs/stdenv/generic/check-meta.nix
index bcb2ca249ddf..a31c02ec225c 100644
--- a/pkgs/stdenv/generic/check-meta.nix
+++ b/pkgs/stdenv/generic/check-meta.nix
@@ -10,7 +10,9 @@ let
     concatMapStrings
     concatMapStringsSep
     concatStrings
+    filter
     findFirst
+    head
     isDerivation
     length
     concatMap
@@ -39,6 +41,16 @@ let
     toPretty
   ;
 
+  unlist = list:
+  if length list == 1
+  then head list
+  else list;
+
+  valOrEmpty = res:
+  if res.success
+  then res.value
+  else [];
+
   # If we're in hydra, we can dispense with the more verbose error
   # messages and make problems easier to spot.
   inHydra = config.inHydra or false;
@@ -303,6 +315,10 @@ let
       (listOf str)
       str
     ];
+    repository = union [
+      (listOf str)
+      str
+    ];
     downloadPage = str;
     changelog = union [
       (listOf str)
@@ -454,8 +470,14 @@ let
     let
       outputs = attrs.outputs or [ "out" ];
       hasOutput = out: builtins.elem out outputs;
-    in
-    {
+    in {
+      repository = valOrEmpty (builtins.tryEval (
+        if attrs ? src.meta.homepage
+        then attrs.src.meta.homepage
+        else if attrs ? srcs && isList attrs.srcs
+        then unlist (map (src: src.meta.homepage) (filter (src: src ? meta.homepage) attrs.srcs))
+        else []));
+
       # `name` derivation attribute includes cross-compilation cruft,
       # is under assert, and is sanitized.
       # Let's have a clean always accessible version here.
```
</details>